Задание 7 - выполнил разбивку по модулям
