--- changeHP.js
@@ -1,9 +1,0 @@
-function changeHP (hp) {
-    this.hp -= hp;
-
-    if (this.hp <= 0) {
-        this.hp = 0;
-    }
-};
-
-export default changeHP;
