+++ modules/const.js
@@ -2,3 +2,11 @@ export const $arenas = document.querySelector('.arenas');
 export const $randomButton = document.querySelector('.button');
 export const $formFight = document.querySelector('.control');
 export const $chat = document.querySelector('.chat');
+
+export const HIT = {
+    head: 30,
+    body: 25,
+    foot: 20,
+};
+
+export const ATTACK = ['head', 'body', 'foot'];
