+++ modules/createPlayer.js
