+++ modules/enemyAttack.js
@@ -1,3 +1,7 @@
+import { getRandom } from './getRandom.js';
+import { ATTACK } from './const.js';
+import { HIT } from './const.js';
+
 function enemyAttack () {
     const hit = ATTACK[getRandom(3) - 1];
     const defence = ATTACK[getRandom(3) - 1];
@@ -8,3 +12,5 @@ function enemyAttack () {
         defence,
     }
 };
+
+export default enemyAttack;
