+++ modules/fightButtonOff.js
@@ -1,3 +1,7 @@
+import { player1,  player2} from './players.js';
+import { $randomButton } from './const.js';
+import createReloadButton from './createReloadButton.js';
+
 function fightButtonOff () {
     if (player1.hp === 0 || player2.hp === 0) {
         $randomButton.disabled = true;
@@ -5,3 +9,5 @@ function fightButtonOff () {
         createReloadButton();
     }
 };
+
+export default fightButtonOff;
