+++ modules/getRandom.js
@@ -1,0 +1,1 @@
+export const getRandom = (meaning) => Math.ceil(Math.random() * meaning);
