+++ modules/playerWins.js
