--- player2.js
@@ -1,16 +1,0 @@
-import elHP from './elHP.js';
-import changeHP from './elHP.js';
-import renderHP from './elHP.js';
-import attack from './elHP.js';
-
-export const player2 = {
-    player: 2,
-    name: 'Sonya',
-    hp: 100,
-    img: 'http://reactmarathon-api.herokuapp.com/assets/sonya.gif',
-    weapon: ['Бамбуковые палки'],
-    elHP,
-    changeHP,
-    renderHP,
-    attack,
-};
