--- renderHP.js
@@ -1,5 +1,0 @@
-function renderHP () {
-    return this.elHP().style.width = this.hp + '%';
-};
-
-export default renderHP;
